PORTFOLIO-T-1 : Implement redux for state management and improved routing
The changes made were to restructure the routing strategy and incorporate redux for managing state. This was done to improve navigation around the application and ensure a unidirectional data flow. The redux store was set up to manage states and corresponding actions, and the `RoutingManager` was introduced to handle navigations. Existing routes were updated to utilize the new systems, including changes to the headerNav function in the NavBar file and the Creation of HomePage function.