Refactor routing procedure and upgrade kvision version
Both the Application State and Routing Manager have been refactored to handle a `Project` enum instead of just an Int for improved functionality and readability. This update also includes an upgrade to the kvisionVersion from 7.0.0 to 7.4.1 in gradle.properties. Lastly, minor changes were made to the formatting of the NavBar links and some console logs were removed for cleaner code.