Add ProjectManagmentSkillPage and update SkillsHomePage
A new file 'ProjectManagmentSkillPage.kt' was added to the project. It outlines various aspects of project management skills. The 'SkillsHomePage.kt' was revised to include a reference to this newly created page. Additionally, minor modifications were made to the 'Card.kt' file to improve its flexibility and applicability to different layouts.